# backend/api/openapi/user.yaml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
  description: User management and role assignment for the Rescue Management System
paths:
  /api/v1/users:
    get:
      summary: List Users
      operationId: listUsers
      tags:
        - Users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            format: int32
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            format: int32
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    format: int32
                  size:
                    type: integer
                    format: int32
                  total:
                    type: integer
                    format: int32
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      summary: Create User
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/v1/users/{id}:
    get:
      summary: Get User
      operationId: getUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update User
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete User
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: User deleted
  /api/v1/roles:
    get:
      summary: List Roles
      operationId: listRoles
      tags:
        - Roles
      responses:
        '200':
          description: List of available roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
  /api/v1/users/{id}/roles:
    put:
      summary: Update User Roles
      operationId: updateUserRoles
      tags:
        - Roles
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roles
              properties:
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: User roles updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        username:
          type: string
        roles:
          type: array
          items:
            type: string
    Role:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
security:
  - bearerAuth: []