# backend/internal/common/genopenapi/user/types.gen.go
// Package user provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package user

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Role defines model for Role.
type Role struct {
	Description *string `json:"description,omitempty"`
	Name        *string `json:"name,omitempty"`
}

// User defines model for User.
type User struct {
	Id       *int32    `json:"id,omitempty"`
	Roles    *[]string `json:"roles,omitempty"`
	Username *string   `json:"username,omitempty"`
}

// ListUsersParams defines parameters for ListUsers.
type ListUsersParams struct {
	Page *int32 `form:"page,omitempty" json:"page,omitempty"`
	Size *int32 `form:"size,omitempty" json:"size,omitempty"`
}

// CreateUserJSONBody defines parameters for CreateUser.
type CreateUserJSONBody struct {
	Password string    `json:"password"`
	Roles    *[]string `json:"roles,omitempty"`
	Username string    `json:"username"`
}

// UpdateUserJSONBody defines parameters for UpdateUser.
type UpdateUserJSONBody struct {
	Password *string   `json:"password,omitempty"`
	Roles    *[]string `json:"roles,omitempty"`
}

// UpdateUserRolesJSONBody defines parameters for UpdateUserRoles.
type UpdateUserRolesJSONBody struct {
	Roles []string `json:"roles"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody CreateUserJSONBody

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody UpdateUserJSONBody

// UpdateUserRolesJSONRequestBody defines body for UpdateUserRoles for application/json ContentType.
type UpdateUserRolesJSONRequestBody UpdateUserRolesJSONBody